+++ source/extensions/omni.isaac.ros2_bridge/docs/CHANGELOG.md
@@ -2,6 +2,13 @@
 CHANGELOG
 **********
 
+[0.8.1] - 2022-05-02
+========================
+
+Changed
+---------
+- Output data type from float to pointf in ROS2 point cloud publisher node
+
 [0.8.0] - 2022-04-29
 ========================